Clean the folder report_sources

This function removes undesirable files and folders present in `report_sources` except for: * `Rmd` files * `README*` files * `cache/` (unless `remove_cache` is `FALSE`)

clean_report_sources(factory = getwd(), quiet = FALSE, remove_cache = TRUE)

Arguments

factory
the path to a report factory; defaults to the current working directory

quiet
a logical indicating if the function should throw a message if content gets removed; defaults to `FALSE`

remove_cache
a logical indicating if `cache` folder should be kept

Value

invisibly returns the path to the removed files

Author

Thibaut Jombart

Examples

## create new random factory in temp folder
odir <- getwd()
random_factory(tempdir())

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpR7iT04/factory_wl4die"

## add content
dir.create(file.path("report_sources", "_archive"))

## check initial content
dir("report_sources", all.files = TRUE)

#> [1] "_archive"                      "."                            
#> [3] ".."                            "example_report_2019-01-31.Rmd"
#> [5] "example_report_2020-11-17.Rmd" "iris_2020-11-15.Rmd"          

## add crap
crap_files <- c("toto.txt",
                "#toto.Rmd#",
                "toto.Rmd~",
                "some_file.html",
                ".hidden")
crap_folders <- c("figures",
                  "cache",
                  "outputs_xlsx")

file.create(file.path("report_sources", crap_files))

#> [1] TRUE TRUE TRUE TRUE TRUE

sapply(file.path("report_sources", crap_folders), dir.create)

#>      report_sources/figures        report_sources/cache 
#>                        TRUE                        TRUE 
#> report_sources/outputs_xlsx 
#>                        TRUE 

## check content with all crap
dir("report_sources", all.files = TRUE)

#>  [1] "_archive"                      "."                            
#>  [3] ".."                            ".hidden"                      
#>  [5] "#toto.Rmd#"                    "cache"                        
#>  [7] "example_report_2019-01-31.Rmd" "example_report_2020-11-17.Rmd"
#>  [9] "figures"                       "iris_2020-11-15.Rmd"          
#> [11] "outputs_xlsx"                  "some_file.html"               
#> [13] "toto.Rmd~"                     "toto.txt"                     

## clean the crap
clean_report_sources()

#> The following files in `report_sources/` are not rmarkdown sources 
#> and were removed:
#> 
#> report_sources/some_file.html
#> report_sources/outputs_xlsx
#> report_sources/#toto.Rmd#
#> report_sources/toto.Rmd~
#> report_sources/toto.txt
#> report_sources/figures
#> report_sources/.hidden
#> report_sources/cache

## check content after cleanup
dir("report_sources", all.files = TRUE)

#> [1] "_archive"                      "."                            
#> [3] ".."                            "example_report_2019-01-31.Rmd"
#> [5] "example_report_2020-11-17.Rmd" "iris_2020-11-15.Rmd"          

setwd(odir)