Collect and ship compiled markdown reports to a single folder

Copies output files for log entries that match given conditions to a dated `shipped_[yyyy-mm-dd]` folder

ship_reports(factory = getwd(), match_exact_type = NULL,
  most_recent = TRUE, outputs_only = FALSE, output_file_types = c(),
  ...)

Arguments

factory
The path the factory directory, defaults to current directory

match_exact_type
a vector of condition types passed to `filter_log` (dots, file, and/or params) that are to be matched EXACTLY to the log entry (no entries with missing components of the specified type will be returned)

most_recent
a logical indicating passed to `filter_log` if only the last log entry should be returned; `FALSE` by default.

outputs_only
a logical indicating if only the outputs of a log entry should be returned passed to `filter_log`; `FALSE` by default.

output_file_types
a vector containing the file types that are to be returned for each entry passed to `filter_log`

...
the arguments that will be used to match and return log entries, which should match the structure of the [compile_reports()] method - passed to `filter_log`

Examples


odir <- getwd()

setwd(tempdir())
random_factory(include_examples = TRUE)

#> [1] "factory_ee10xomb92Q9Q"

source_name <- "foo"
report_source_file_name <- list_reports(pattern = source_name)[1]

dots_args <- list("lots" = data.frame(a = c(10,20)))
compile_report(
  report_source_file_name,
  quiet = FALSE,
  params = list("other" = "two",
                "more" = list("thing" = "foo")),
  extra = dots_args)

#> 
#> /// compiling report: 'foo_2018-06-29'

#> // using params: 
#> other: two
#> more: foo

#> 
#> 
#> processing file: foo_2018-06-29.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.....                                                                 |   7%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........                                                             |  13%
#> label: options (with options) 
#> List of 4
#>  $ include: logi FALSE
#>  $ message: logi FALSE
#>  $ warning: logi FALSE
#>  $ error  : logi FALSE
#> 
#>   |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................                                                   |  27%
#> label: iris
#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: boxplots

#>   |                                                                              |.................................                                     |  47%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................                                 |  53%
#> label: violins

#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: summary_base
#>   |                                                                              |...................................................                   |  73%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: summary_tidy
#>   |                                                                              |.............................................................         |  87%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................................................................     |  93%
#> label: outputs
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: foo_2018-06-29.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS foo_2018-06-29.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output foo_2018-06-29.html --email-obfuscation none --self-contained --standalone --section-divs --template "C:\Users\steph\OneDrive\Documents\R\win-library\3.6\rmarkdown\rmd\h\default.html" --no-highlight --variable highlightjs=1 --variable "theme:bootstrap" --include-in-header "C:\Users\steph\AppData\Local\Temp\RtmpMjWkL4\rmarkdown-str14d41067245a.html" --mathjax --variable "mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" --lua-filter "C:/Users/steph/OneDrive/Documents/R/win-library/3.6/rmarkdown/rmd/lua/pagebreak.lua" --lua-filter "C:/Users/steph/OneDrive/Documents/R/win-library/3.6/rmarkdown/rmd/lua/latex-div.lua" 

#> 
#> Output created: foo_2018-06-29.html

#> 
#> /// 'foo_2018-06-29' done!

#> Warning: cannot remove file 'figures', reason 'Permission denied'


compile_report(
  report_source_file_name,
  quiet = TRUE,
  params = list(other = "test"))

#> 
#> /// compiling report: 'foo_2018-06-29'

#> // using params: 
#> other: test

#> 
#> /// 'foo_2018-06-29' done!


compile_report(
  report_source_file_name,
  quiet = FALSE,
  params = list("other" = "two",
                "more" = list("thing" = "foo")),
  extra = dots_args)

#> 
#> /// compiling report: 'foo_2018-06-29'

#> // using params: 
#> other: two
#> more: foo

#> 
#> 
#> processing file: foo_2018-06-29.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.....                                                                 |   7%
#>   ordinary text without R code
#> 
#>   |                                                                              |.........                                                             |  13%
#> label: options (with options) 
#> List of 4
#>  $ include: logi FALSE
#>  $ message: logi FALSE
#>  $ warning: logi FALSE
#>  $ error  : logi FALSE
#> 
#>   |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................                                                   |  27%
#> label: iris
#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: boxplots

#>   |                                                                              |.................................                                     |  47%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................                                 |  53%
#> label: violins

#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: summary_base
#>   |                                                                              |...................................................                   |  73%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: summary_tidy
#>   |                                                                              |.............................................................         |  87%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................................................................     |  93%
#> label: outputs
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: foo_2018-06-29.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS foo_2018-06-29.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output foo_2018-06-29.html --email-obfuscation none --self-contained --standalone --section-divs --template "C:\Users\steph\OneDrive\Documents\R\win-library\3.6\rmarkdown\rmd\h\default.html" --no-highlight --variable highlightjs=1 --variable "theme:bootstrap" --include-in-header "C:\Users\steph\AppData\Local\Temp\RtmpMjWkL4\rmarkdown-str14d4741e216d.html" --mathjax --variable "mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" --lua-filter "C:/Users/steph/OneDrive/Documents/R/win-library/3.6/rmarkdown/rmd/lua/pagebreak.lua" --lua-filter "C:/Users/steph/OneDrive/Documents/R/win-library/3.6/rmarkdown/rmd/lua/latex-div.lua" 

#> 
#> Output created: foo_2018-06-29.html

#> 
#> /// 'foo_2018-06-29' done!


log_file <- readRDS(".compile_log.rds")

ship_reports(
  params = list("other" = "two"),
  most_recent = TRUE
)

#> 
#> /// Shipping foo_2018-06-29 outputs: 
#> 
#> boxplots-1.pdf
#> boxplots-1.png
#> violins-1.pdf
#> violins-1.png
#> foo_2018-06-29.html
#> outputs_base.csv


setwd(odir)