Inspect and validate the content of a factory

validate_factory() can be used to inspect the content of a factory and make everything looks fine. This includes various sanity checks listed in details that will error if a problem is found.

validate_factory(factory = ".")

Arguments

factory
The path to the report factory or a folder within the desired factory. Defaults to the current directory.

Value

A list with 4 entries:

root - the root folder path of the factory;

factory_name - the name of the report factory;

report_sources - the name of the report_sources folder; and

outputs - the name of the outputs folder.

Details

Checks run on the factory include:

the factory directory exists;

the factory_config file exist;

all mandatory folders exist - by default these are 'report_sources/' and 'outputs/';