Filter reportfactory logs by factory, source file name, parameters, dots arguments, output file types, or other logged data

By default the function returns all matches --- even if some conditions are missing, or there are other log conditions present

filter_log(log_file, match_exact_type = NULL, most_recent = FALSE,
  outputs_only = FALSE, output_file_types = c(), ...)

Arguments

log_file
the compile log list (which is located in the root of the factory and saved as a .rds file)

match_exact_type
a vector of condition types (dots, file, and/or params) that are to be matched EXACTLY to the log entry (no entries with missing components of the specified type will be returned)

most_recent
a logical indicating if only the last log entry should be returned; `FALSE` by default.

outputs_only
a logical indicating if only the outputs of a log entry should be returned; `FALSE` by default.

output_file_types
a vector containing the file types that are to be returned for each entry

...
the arguments that will be used to match and return log entries, which should match the structure of the `reportfactory::compile_reports` method

Value

a list of log entries which match the `...` arguments and other parameter input

Examples

## This filters a compile log with a set of conditions, which would 
## match the call of the `reportfactory::compile_report` function. 

## new random factory in temp folder
odir <- getwd()
new_factory(tempdir(), include_examples = TRUE)

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpZUTLyi"

factory_name = "foo"
report_source_file_name <- list_reports(pattern = factory_name)[1]

## The log entry for a report created with the following arguments:
compile_report(file = report_source_file_name,
               quiet = TRUE,
               params = list(other = "test"))

#> 
#> /// compiling report: 'foo_2018-06-29'

#> // using params: 
#> other: test

#> 
#> /// 'foo_2018-06-29' done!

## The log is saved in the root of the factory
log_file <- readRDS(".compile_log.rds")

## Filter the log file for the entry with the same arguments:
filter_log(log_file = log_file,
           file = report_source_file_name,
           quiet = TRUE,
           params = list(other = "test"))

#> $`2020-03-17_19-32-18`
#> $`2020-03-17_19-32-18`$compile_init_env
#> $`2020-03-17_19-32-18`$compile_init_env$file
#> [1] "foo_2018-06-29.Rmd"
#> 
#> $`2020-03-17_19-32-18`$compile_init_env$quiet
#> [1] TRUE
#> 
#> $`2020-03-17_19-32-18`$compile_init_env$factory
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi"
#> 
#> $`2020-03-17_19-32-18`$compile_init_env$clean_report_sources
#> [1] FALSE
#> 
#> $`2020-03-17_19-32-18`$compile_init_env$remove_cache
#> [1] TRUE
#> 
#> $`2020-03-17_19-32-18`$compile_init_env$encoding
#> [1] "UTF-8"
#> 
#> $`2020-03-17_19-32-18`$compile_init_env$params
#> $`2020-03-17_19-32-18`$compile_init_env$params$other
#> [1] "test"
#> 
#> 
#> 
#> $`2020-03-17_19-32-18`$dots
#> list()
#> 
#> $`2020-03-17_19-32-18`$timestamp
#> [1] "2020-03-17_19:32:18"
#> 
#> $`2020-03-17_19-32-18`$output_dir
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi/report_outputs/foo_2018-06-29/compiled_other_test_2020-03-17_19-32-18"
#> 
#> $`2020-03-17_19-32-18`$output_files
#> $`2020-03-17_19-32-18`$output_files[[1]]
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi/report_outputs/foo_2018-06-29/compiled_other_test_2020-03-17_19-32-18/figures/boxplots-1.pdf"
#> 
#> $`2020-03-17_19-32-18`$output_files[[2]]
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi/report_outputs/foo_2018-06-29/compiled_other_test_2020-03-17_19-32-18/figures/boxplots-1.png"
#> 
#> $`2020-03-17_19-32-18`$output_files[[3]]
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi/report_outputs/foo_2018-06-29/compiled_other_test_2020-03-17_19-32-18/figures/violins-1.pdf"
#> 
#> $`2020-03-17_19-32-18`$output_files[[4]]
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi/report_outputs/foo_2018-06-29/compiled_other_test_2020-03-17_19-32-18/figures/violins-1.png"
#> 
#> $`2020-03-17_19-32-18`$output_files[[5]]
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi/report_outputs/foo_2018-06-29/compiled_other_test_2020-03-17_19-32-18/foo_2018-06-29.html"
#> 
#> $`2020-03-17_19-32-18`$output_files[[6]]
#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpZUTLyi/report_outputs/foo_2018-06-29/compiled_other_test_2020-03-17_19-32-18/outputs_base.csv"
#> 
#> 
#> 

setwd(odir)