Create a new report factory

This function can be used to create a new report factory, using a template containing data and associated reports as examples. The provided examples use (simulated) infectious disease data, but any other data analysis would work the same. See details for the content of the factory

new_factory(destination = "new_factory", include_template = TRUE,
  include_examples = FALSE, move_in = TRUE)

Arguments

destination
the name of the report factory folder to be created

include_template
a logical indicating if a template of report and folders structure shoud be added to the factory; defaults to TRUE

include_examples
a logical indicating if examples of reports shoud be added to the factory; defaults to FALSE

move_in
a logical indicating if the current session should move into the created factory; defaults to TRUE

Details

The default factory includes:

data/: a folder storing data

report_sources/: a folder storing the reports, named after the convention described in update_reports, and stored in subfolders contacts and epicurves

.gitignore: a file used to tell git to ignore the produced outputs in report_outputs

open.Rproj: an Rproject to open the factory using Rstudio

.here: an empty file used as anchor by here

Examples

if (FALSE) {
destination <- file.path(tempdir(), "new_factory")
destination
new_factory(destination)
dir()

## check content
list_reports()
list_outputs()

## check dependencies
list_deps()
install_deps()

## compile a single report:

compile_report("contacts_2017-10-29", quiet = TRUE)
list_outputs()

## compile all reports (only most recent versions):

update_reports()
list_outputs()
}