Inspect and validate the content of a factory

This function can be used to inspect the content of a factory and make sure it looks fine. This includes various sanity checks listed in details. The function returns a list of error and warning messages.

validate_factory(factory = getwd(), warnings = TRUE, errors = TRUE)

Arguments

factory
the path to a report factory; defaults to the current working directory

warnings
A logical indicating if warnings should be issued; defaults to TRUE.

errors
A logical indicating if errors should be issued; defaults to TRUE.

Details

Checks ran on the factory include (the result of a failure is indicated in brackets):

the directory exists [error]

all mandatory files exist: .here, .gitignore [error]

all mandatory folders exist: report_sources/ [error]

all .Rmd reports have unique names once outside their folders (to avoid conflicts in outputs) [warning]